refactor: address review shared LocalisedStringSchema
