+++ client/public/mocks/analiza-i-struktury-danych.json
