+++ client/src/hooks/useCourse.ts
@@ -1,13 +1,16 @@
 import { useQuery } from '@tanstack/react-query';
-import axios from 'axios';
-import { CourseSchema, type Course } from '@/types/course';
+import { CourseSchema, type Course } from '@/schemas/course';
+import { agent } from '@/lib/api';
 
-export default function useCourse(slug: string) {
+async function fetchCourse(slug: string): Promise<Course> {
+  const response = await agent.get(`/mocks/${slug}.json`);
+  return CourseSchema.parse(response.data);
+}
+
+export function useCourse(slug: string) {
   return useQuery<Course>({
     queryKey: ['course', slug],
-    queryFn: async () => {
-      const response = await axios.get('/mocks/course.json');
-      return CourseSchema.parse(response.data);
-    },
+    queryFn: () => fetchCourse(slug),
+    staleTime: 1000 * 60 * 5, // 5 minutes
   });
 }
